refactor(whispering): collapse focus-main + navigate into revealMainWindow
Two halves of one concept lived apart: the recording overlay's
focus-main event revealed the main window but never navigated, and the
picker's navigate-main-window event navigated but never revealed. So
clicking "Manage transformations" in the picker routed the main window to
/transformations without bringing it forward, leaving a minimized main
window navigated where the user couldn't see it.

Collapse both into one revealMainWindow({ path? }) channel with a single
listener owner, attachMainWindowReveal, that reveals first (show +
unminimize + focus) and then routes if a path was given. Delete
attach-deep-link-navigation.ts (whose name and docstring claimed
deep-link/tray duties it never had) and the focus-main listener, so
attach-recording-overlay goes back to pure overlay status/action.

Overlay-pill behavior is unchanged (empty payload -> reveal only); the
picker now reveals as well as routes, fixing the bug above.
