+++ apps/whispering/src/routes/(app)/_runtime/attach-recording-overlay.svelte.ts
@@ -1,5 +1,4 @@
 import type { UnlistenFn } from '@tauri-apps/api/event';
-import { getCurrentWindow } from '@tauri-apps/api/window';
 import { recordingOverlay } from '#platform/recording-overlay';
 import { tauri } from '#platform/tauri';
 import {
@@ -10,14 +9,12 @@ import {
 import {
 	type RecordingOverlayStatus,
 	recordingOverlayAction,
-	recordingOverlayFocusMain,
 } from '$lib/recording-overlay/events';
 import { manualRecorder } from '$lib/state/manual-recorder.svelte';
 import { vadRecorder } from '$lib/state/vad-recorder.svelte';
 
 export function attachRecordingOverlay() {
 	let unlistenAction: UnlistenFn | undefined;
-	let unlistenFocus: UnlistenFn | undefined;
 
 	const overlayStatus = $derived.by((): RecordingOverlayStatus | null => {
 		if (manualRecorder.state === 'RECORDING')
@@ -45,19 +42,10 @@ export function attachRecordingOverlay() {
 				}
 				if (event.payload === 'stop') void stopVadRecording();
 			});
-			unlistenFocus = await recordingOverlayFocusMain.listen(() => {
-				const mainWindow = getCurrentWindow();
-				void (async () => {
-					await mainWindow.show();
-					await mainWindow.unminimize();
-					await mainWindow.setFocus().catch(() => {});
-				})();
-			});
 		})();
 	}
 
 	return () => {
 		unlistenAction?.();
-		unlistenFocus?.();
 	};
 }
