+++ apps/whispering/src/routes/(app)/_runtime/runtime-owners.ts
@@ -2,9 +2,9 @@ import { dictationCapability } from '$lib/state/dictation-capability.svelte';
 import { attachAnalytics } from './attach-analytics.svelte';
 import { attachAutoPasteIntent } from './attach-auto-paste-intent.svelte';
 import { attachDebugCommands } from './attach-debug-commands';
-import { attachDeepLinkNavigation } from './attach-deep-link-navigation';
 import { attachGlobalShortcutTriggers } from './attach-global-shortcut-triggers';
 import { attachLocalShortcutListener } from './attach-local-shortcut-listener.svelte';
+import { attachMainWindowReveal } from './attach-main-window-reveal';
 import { attachRecordingOverlay } from './attach-recording-overlay.svelte';
 import { attachRecordingRetention } from './attach-recording-retention.svelte';
 import { attachShortcutSync } from './attach-shortcut-sync';
@@ -24,7 +24,7 @@ export const runtimeOwners = [
 	{ attach: attachUnloadPolicy },
 	{ attach: attachRecordingRetention },
 	{ attach: attachUpdateCheck },
-	{ attach: attachDeepLinkNavigation },
+	{ attach: attachMainWindowReveal },
 	{ attach: attachAutoPasteIntent },
 	{ attach: dictationCapability.attach },
 ] satisfies RuntimeOwner[];
