+++ apps/whispering/src/routes/recording-overlay/+page.svelte
@@ -5,10 +5,10 @@
 	import XIcon from '@lucide/svelte/icons/x';
 	import { type UnlistenFn } from '@tauri-apps/api/event';
 	import { onDestroy, onMount } from 'svelte';
+	import { revealMainWindow } from '$lib/main-window';
 	import {
 		type RecordingOverlayAction,
 		recordingOverlayAction,
-		recordingOverlayFocusMain,
 		recordingOverlayMicLevel,
 		recordingOverlayReady,
 		recordingOverlayStatus,
@@ -75,7 +75,7 @@
 	}
 
 	function focusMainWindow() {
-		void recordingOverlayFocusMain.emit();
+		void revealMainWindow.emit({});
 	}
 </script>
